src/main/java/com/rest/documentManager/service/impl/AttachmentServiceImpl.java
package com.rest.documentManager.service.impl;

import com.rest.documentManager.entity.Attachment;
import com.rest.documentManager.repository.AttachmentRepository;
import com.rest.documentManager.service.AttachmentService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.util.StringUtils;
import org.springframework.web.multipart.MultipartFile;

@Service
public class AttachmentServiceImpl implements AttachmentService {

    @Autowired
    private AttachmentRepository attachmentRepository;

    @Override
    public Attachment saveAttachment(MultipartFile file) throws Exception {
        String fileName = StringUtils.cleanPath(file.getOriginalFilename());

        try {
            if (fileName.contains("..")) {
                throw new Exception("Filename contains invalid path sequence " + fileName);
            }

            Attachment attachment = new Attachment(fileName, file.getContentType(), file.getBytes());
            return attachmentRepository.save(attachment);
        } catch (Exception err) {
            throw new Exception("Could not save File: " + fileName);
        }
    }

    @Override
    public Attachment getAttachment(Integer id) throws Exception {
        return attachmentRepository.findById(id).orElseThrow(() -> new Exception("File not found with Id: " + id));
    }
}
